Added command buffer lists

+++ VulkanSimplifiedTest/SwapchainSizes.h
@@ -3,4 +3,4 @@
 constexpr uint32_t width = 1920U;
 constexpr uint32_t height = 1080U;
 
-constexpr size_t framesInFlight = 3U;
+constexpr uint32_t framesInFlight = 3U;
